move queries, update tests
